// frontend/src/shared/types/global.types.ts
export interface ApiResponse<T> {
  success: boolean;
  message: string | null;
  data: T | null;
  errorCode: string | null;
  timestamp: string;
}

export interface PageResponse<T> {
  content: T[];
  totalElements: number;
  totalPages: number;
  size: number;
  number: number;
  first: boolean;
  last: boolean;
}